refactor sourcing plan command DTOs
